[Refactor] update advisors to use builder pattern and adjust import statements for ToolUtils
