+++ spring-ai-alibaba-core/src/test/java/com/alibaba/cloud/ai/dashscope/chat/DashScopeMultiModalChatTests.java
@@ -20,6 +20,7 @@ import java.net.URI;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
 
 import com.alibaba.cloud.ai.dashscope.api.DashScopeApi;
@@ -119,10 +120,10 @@ public class DashScopeMultiModalChatTests {
 
 		// Create media list with URL
 		List<Media> mediaList = List.of(new Media(MimeTypeUtils.IMAGE_PNG,
-				new URI("https://dashscope.oss-cn-beijing.aliyuncs.com/images/dog_and_girl.jpeg").toURL()));
+				new URI("https://dashscope.oss-cn-beijing.aliyuncs.com/images/dog_and_girl.jpeg").toURL().toURI()));
 
 		// Create user message with media
-		UserMessage message = new UserMessage(TEST_PROMPT, mediaList);
+		UserMessage message = UserMessage.builder().media(mediaList).text(TEST_PROMPT).build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.IMAGE);
 
 		// Create prompt with options
@@ -154,8 +155,10 @@ public class DashScopeMultiModalChatTests {
 		when(dashScopeApi.chatCompletionEntity(any(ChatCompletionRequest.class))).thenReturn(responseEntity);
 
 		// Create user message with resource media
-		UserMessage message = new UserMessage(TEST_PROMPT,
-				new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")));
+		UserMessage message = UserMessage.builder()
+			.text(TEST_PROMPT)
+			.media(new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")))
+			.build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.IMAGE);
 
 		// Create prompt with options
@@ -193,7 +196,7 @@ public class DashScopeMultiModalChatTests {
 		}
 
 		// Create user message with media
-		UserMessage message = new UserMessage(TEST_VIDEO_PROMPT, mediaList);
+		UserMessage message = UserMessage.builder().media(mediaList).text(TEST_VIDEO_PROMPT).build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.VIDEO);
 
 		// Create prompt with options
@@ -230,8 +233,10 @@ public class DashScopeMultiModalChatTests {
 		when(dashScopeApi.chatCompletionStream(any(ChatCompletionRequest.class))).thenReturn(Flux.just(chunk1, chunk2));
 
 		// Create user message with resource media
-		UserMessage message = new UserMessage(TEST_PROMPT,
-				new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")));
+		UserMessage message = UserMessage.builder()
+			.text(TEST_PROMPT)
+			.media(new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")))
+			.build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.IMAGE);
 
 		// Create prompt with options
@@ -269,10 +274,10 @@ public class DashScopeMultiModalChatTests {
 
 		// Create media list with URL
 		List<Media> mediaList = List.of(new Media(MimeTypeUtils.IMAGE_PNG,
-				new URI("https://dashscope.oss-cn-beijing.aliyuncs.com/images/dog_and_girl.jpeg").toURL()));
+				new URI("https://dashscope.oss-cn-beijing.aliyuncs.com/images/dog_and_girl.jpeg").toURL().toURI()));
 
 		// Create user message with media
-		UserMessage message = new UserMessage(TEST_PROMPT, mediaList);
+		UserMessage message = UserMessage.builder().media(mediaList).text(TEST_PROMPT).build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.IMAGE);
 
 		// Create prompt
@@ -304,8 +309,10 @@ public class DashScopeMultiModalChatTests {
 		DashScopeChatModel realChatModel = DashScopeChatModel.builder().dashScopeApi(realApi).build();
 
 		// Create user message with resource media
-		UserMessage message = new UserMessage(TEST_PROMPT,
-				new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")));
+		UserMessage message = UserMessage.builder()
+			.text(TEST_PROMPT)
+			.media(new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")))
+			.build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.IMAGE);
 
 		// Create prompt
@@ -343,7 +350,7 @@ public class DashScopeMultiModalChatTests {
 		}
 
 		// Create user message with media
-		UserMessage message = new UserMessage(TEST_VIDEO_PROMPT, mediaList);
+		UserMessage message = UserMessage.builder().media(mediaList).text(TEST_VIDEO_PROMPT).build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.VIDEO);
 
 		// Create prompt
@@ -375,8 +382,10 @@ public class DashScopeMultiModalChatTests {
 		DashScopeChatModel realChatModel = DashScopeChatModel.builder().dashScopeApi(realApi).build();
 
 		// Create user message with resource media
-		UserMessage message = new UserMessage(TEST_PROMPT,
-				new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")));
+		UserMessage message = UserMessage.builder()
+			.text(TEST_PROMPT)
+			.media(new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")))
+			.build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.IMAGE);
 
 		// Create prompt
@@ -418,8 +427,10 @@ public class DashScopeMultiModalChatTests {
 		DashScopeChatModel realChatModel = DashScopeChatModel.builder().dashScopeApi(realApi).build();
 
 		// Create user message with resource media and custom prompt
-		UserMessage message = new UserMessage("请详细描述这张图片中的场景，包括人物、动物、环境等细节，并分析图片的情感基调。",
-				new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")));
+		UserMessage message = UserMessage.builder()
+			.media(new Media(MimeTypeUtils.IMAGE_JPEG, new ClassPathResource("multimodel/dog_and_girl.jpeg")))
+			.text("请详细描述这张图片中的场景，包括人物、动物、环境等细节，并分析图片的情感基调。")
+			.build();
 		message.getMetadata().put(MESSAGE_FORMAT, MessageFormat.IMAGE);
 
 		// Create prompt
