+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-example/src/main/java/com/alibaba/cloud/ai/example/graph/bigtool/agent/CalculateAgent.java
@@ -22,8 +22,11 @@ import org.springframework.ai.chat.client.ChatClient;
 import org.springframework.ai.chat.model.ChatResponse;
 import org.springframework.ai.document.Document;
 import org.springframework.ai.tool.ToolCallback;
+import org.springframework.ai.tool.definition.DefaultToolDefinition;
 import org.springframework.ai.tool.definition.ToolDefinition;
 import org.springframework.ai.tool.method.MethodToolCallback;
+import org.springframework.ai.tool.support.ToolUtils;
+import org.springframework.ai.util.json.schema.JsonSchemaGenerator;
 import org.springframework.util.ReflectionUtils;
 import org.springframework.util.StringUtils;
 
@@ -66,8 +69,13 @@ public class CalculateAgent implements NodeAction {
 			var toolMethod = ReflectionUtils.findMethod(Math.class, document.getMetadata().get(METHOD_NAME).toString(),
 					(Class<?>[]) document.getMetadata().get(METHOD_PARAMETER_TYPES));
 
+			DefaultToolDefinition.Builder toolDefinitionBuilder = DefaultToolDefinition.builder()
+				.name(ToolUtils.getToolName(toolMethod))
+				.description(ToolUtils.getToolDescription(toolMethod))
+				.inputSchema(JsonSchemaGenerator.generateForMethodInput(toolMethod));
+
 			MethodToolCallback build = MethodToolCallback.builder()
-				.toolDefinition(ToolDefinition.builder(toolMethod).description(document.getText()).build())
+				.toolDefinition(toolDefinitionBuilder.description(document.getText()).build())
 				.toolMethod(toolMethod)
 				.build();
