+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-example/src/main/java/com/alibaba/cloud/ai/example/graph/openmanus/tool/PythonExecute.java
@@ -22,7 +22,6 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.TypeReference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.ai.chat.memory.InMemoryChatMemory;
 import org.springframework.ai.chat.model.ToolContext;
 import org.springframework.ai.openai.api.OpenAiApi;
 import org.springframework.ai.tool.annotation.ToolParam;
@@ -75,8 +74,6 @@ public class PythonExecute implements BiFunction<String, ToolContext, ToolExecut
 			.inputType(String.class)
 			.build();
 	}
-
-	private InMemoryChatMemory chatMemory;
 
 	public ToolExecuteResult run(String toolInput) {
 		log.info("PythonExecute toolInput:{}", toolInput);
