+++ spring-ai-alibaba-jmanus/src/main/java/com/alibaba/cloud/ai/example/manus/llm/LlmService.java
@@ -352,14 +352,15 @@ public class LlmService {
 			.build();
 
 		// // 每个agent执行过程中，用独立的memroy
+
 		this.agentExecutionClient = ChatClient.builder(chatModel)
-			.defaultAdvisors(new MessageChatMemoryAdvisor(agentMemory))
+			.defaultAdvisors(MessageChatMemoryAdvisor.builder(agentMemory).build())
 			.defaultAdvisors(new SimpleLoggerAdvisor())
 			.defaultOptions(OpenAiChatOptions.builder().internalToolExecutionEnabled(false).build())
 			.build();
 
 		this.finalizeChatClient = ChatClient.builder(chatModel)
-			.defaultAdvisors(new MessageChatMemoryAdvisor(conversationMemory))
+			.defaultAdvisors(MessageChatMemoryAdvisor.builder(conversationMemory).build())
 			.defaultAdvisors(new SimpleLoggerAdvisor())
 			.build();
