+++ spring-ai-alibaba-studio/src/main/java/com/alibaba/cloud/ai/service/impl/ChatClientDelegateImpl.java
@@ -35,15 +35,12 @@ import java.util.UUID;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.ai.chat.client.DefaultChatClient;
-import org.springframework.ai.chat.client.advisor.AbstractChatMemoryAdvisor;
-import org.springframework.ai.chat.client.advisor.api.Advisor;
 import org.springframework.ai.chat.model.ChatModel;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ReflectionUtils;
 import org.springframework.util.StringUtils;
 
-import static org.springframework.ai.chat.client.advisor.AbstractChatMemoryAdvisor.CHAT_MEMORY_CONVERSATION_ID_KEY;
-import static org.springframework.ai.chat.client.advisor.AbstractChatMemoryAdvisor.CHAT_MEMORY_RETRIEVE_SIZE_KEY;
+import static org.springframework.ai.chat.memory.ChatMemory.CONVERSATION_ID;
 
 @Service
 @Slf4j
@@ -101,8 +98,8 @@ public class ChatClientDelegateImpl implements ChatClientDelegate {
 			chatID = UUID.randomUUID().toString();
 		}
 		String finalChatID = chatID;
-		clientRequestSpec.advisors(spec -> spec.param(CHAT_MEMORY_CONVERSATION_ID_KEY, finalChatID)
-			.param(CHAT_MEMORY_RETRIEVE_SIZE_KEY, CHAT_MEMORY_RETRIEVE_SIZE));
+		clientRequestSpec.advisors(spec -> spec.param(CONVERSATION_ID, finalChatID)
+			.param("chat_memory_response_size", CHAT_MEMORY_RETRIEVE_SIZE));
 
 		String resp = clientRequestSpec.user(input).call().content();
 		return ChatClientRunResult.builder()
@@ -136,15 +133,16 @@ public class ChatClientDelegateImpl implements ChatClientDelegate {
 					// todo 扩展其他项
 
 					// 获取是否开启memory
-					for (Advisor advisor : defaultChatClientRequest.getAdvisors()) {
-						try {
-							Class<?> clazz = advisor.getClass();
-							client.setIsMemoryEnabled(AbstractChatMemoryAdvisor.class.isAssignableFrom(clazz));
-						}
-						catch (Exception e) {
-							client.setIsMemoryEnabled(false);
-						}
-					}
+					// TODO AbstractChatMemoryAdvisor 类已删除，需优化该部分
+					// for (Advisor advisor : defaultChatClientRequest.getAdvisors()) {
+					// try {
+					// Class<?> clazz = advisor.getClass();
+					// client.setIsMemoryEnabled(AbstractChatMemoryAdvisor.class.isAssignableFrom(clazz));
+					// }
+					// catch (Exception e) {
+					// client.setIsMemoryEnabled(false);
+					// }
+					// }
 
 					// 获取chatModel并设置
 					// 目前仅支持 ModelType.CHAT 类型，暂不支持其他类型的Model
